Bug 1
Added option (in scodegen) to delete rows before BatchInsert

setting this option in GCOMSOrderStatus on a cold restart, overrides
cache entry by deleting rows in that case.  Also will delete rows when
creating ordr_status entries for new orders, but I don't think this
delay is on the critical path to anything.

This appears to have fixed the bug of the conflicting primary keys.