Bug 84
wrote test program and re-wrote the subsecond formatting, and tested it.

Basic description of bug and its solution:

It's a common practice in timekeeping applications NOT to round
values off.  Instead we truncate.  For instance, June 29th is
considered to be "June" not "July" even though rounding off the
date to the nearest month would yield "July".   Same here.  A time
of 14:59:59.9999, stated to the nearest second, is 14:59:59, NOT
15:00:00.

Modula-3's standard formatting routines don't have a setting for truncating
floating-point numbers like this, so I had to write my own.  A new restriction
on TZ.FormatSubsecond is that it can't format finer than a microsecond.  But
this restriction can be removed in future, if necessary.