Bug 85
This is the checkin of the proposed fix to the problem mentioned in Comment 9 .

The problem was with running a child process that prints output to standard out (or error).

A "ForkWriter" is run, conditionally, for each of std. output and std. error.

On a Wait, the implementation waited only for the child process to exit.  However, in order
for Wait to return, all the outputs must have been collected, which means that the two
Writers must also have exited.

The fix adds Thread.Joins for the two Writers.  It depends critically on the fact that the Writers
exit at some time after the child process exits.