Bug 63
The bug appears fixed, finally.

Added routines so that the interfaces and implementations are first
generated with the extension TMP.  If the contents of the TMP file are
the same as that of the final (non-TMP) file, the files are not renamed to the final filename.

required updating SchemeM3 because I dont want sstubgen to depend on
bootstrapping with itself, which it would if we called directly into
Modula-3 from the Scheme code of sstubgen itself.