Description of change: (what change is being made?) Enable automation of RDL analysis and generation steps, following GMF rubric
Why is the change being made?
Madison Bay will use RDL as its register description format, and generate
as much collateral as practical.

Who else has participated in reviewing the change?
Confirmed with Dan Hendrickson, Justin Baker and others that they are okay with
architecture team doing this level of automation with the repo.

HSD: N/A