Add mby_server and shared lib libmbymodel.so
Add a MBY model server based on the HLP implementation and socket
interface. To compile:

  make mby_server

This will also build the shared library libmbymodel.so where the actual
implementation of the pipeline exists.
The file mby_model.h contains the public API that can be used to control
the model from the server.

Change-Id: I6a83877b63ffc08aa216ee9a559f2c02676769b3
Reviewed-on: https://git-amr-3.devtools.intel.com/gerrit/185857
Tested-by: sys_build <sys_build@intel.com>
Reviewed-by: Nagler, Thomas J <thomas.j.nagler@intel.com>