Exclude Scala build from GK checker
The Scala build often fails because of infrastructure issues.
In particular, we saw failures with the download of 3rd party
dependencies that are not connected with the changes to Scala.
This commit exclude the Scala from the default GK build since these
failures are blocking the turnins of non-Scala changes.

Change-Id: I9bf5dcf1fe59c90d4004ae2e1cc25c48396951a2