fixed concurrency bug and improved debug output
Analysis:

Symptom:

Trace-conversion program launches a bunch of threads.

Commands threads to exit.

Threads claim they have exited. [I just figured this out today, actually...]

Master process hangs and will not exit.  [Long-annoying symptom]

The code:

Here mu is a mutex and d is the condition variable used for assignment of jobs.

worker.freeP() returns TRUE if and only if the worker is idle and ready to accept new work.

I. Job assignment:

      WHILE NOT assigned DO
        FOR w := FIRST(workers^) TO LAST(workers^) DO
          IF workers[w].freeP() THEN
            IF doDebug THEN
              Debug.Out(F("Fsdb.Parse : assigning partial trace file %s to worker %s", Int(i), Int(w)));
            END;

            workers[w].task(wdWr[i], fileTab[i], wdPth[i]);
            assigned := TRUE;
            EXIT
          END
        END;
        IF NOT assigned THEN
          LOCK mu DO
            Thread.Wait(mu, d)
          END
        END
      END

II. Shutdown wait:

     (* wait for workers to be completely done *)
     FOR i := FIRST(workers^) TO LAST(workers^) DO
       (* wait for workers to finish *)
      WHILE NOT workers[i].freeP() DO
        LOCK mu DO
          Thread.Wait(mu, d)
        END
       END
     END

Both parts of the code are broken.  I found it more obvious in the second part.

If a worker w is free, the corresponding predicate "w.freeP()" is stable, it will not change until you give a job to the worker.  So I had reasoned that it is perfectly OK to check it outside the critical section: yes, it is true that you might miss the free-ness of the worker, but so what?  It might have waited a bit longer.  (The reason I wanted to check it outside of the critical section is that I was too lazy to add a variable for which worker to assign to: the workers are doing other things under mutexes, so I wanted to avoid any deadlocks that might result from actually assigning a job inside the critical section.  I'm not exactly sure what worker.task() might do, but I suspect it will try to acquire the mu mutex itself---which would lead to a deadlock if called from within the C.S.)

But if a worker is non-free, that is, working, the predicate "NOT w.freeP()" is unstable.  I can't check that outside of a critical section and take a decision based on that evaluation.  It might be wrong by the time I take the decision.  In particular, in the shutdown wait, if the worker became free (and marked itself so) right after my check for free-nees and before I acquire the mutex, I will miss that the worker is free.  The worker will not signal more than once, so the system deadlocks with me waiting for a signal that never arrives!

In the first code it's the same.  The safety part is satisfied---the job assignment is OK, I will never assign a job to a busy worker, becase freeP() is indeed stable.

But the second half is "IF NOT assigned" which is effectively  worker[0].freeP() OR worker[1].freeP() OR .. OR worker[N-1].freeP() and *that* certainly isn't stable!

I probably won't notice it in the task-assignment code, though, because there are many threads running and signalling on d, so even if I miss a signal, another one comes along later for something else and wakes up the system.

But in the shutdown code, the *last* worker to exit may miss the mutex, and bang, deadlock...no one else will signal after the last exiter.

See checkin for fix